+++ api/tests/Feature/AAUserTest.php
@@ -1,0 +1,395 @@
+<?php
+
+namespace Tests\Feature;
+
+use App\Enums\PoolCandidateStatus;
+use App\Facades\Notify;
+use App\Models\Pool;
+use App\Models\PoolCandidate;
+use App\Models\User;
+use Database\Seeders\ClassificationSeeder;
+use Database\Seeders\RolePermissionSeeder;
+use Illuminate\Foundation\Testing\RefreshDatabase;
+use Nuwave\Lighthouse\Testing\MakesGraphQLRequests;
+use Nuwave\Lighthouse\Testing\RefreshesSchemaCache;
+use Tests\TestCase;
+use Tests\UsesProtectedGraphqlEndpoint;
+
+class AAUserTest extends TestCase
+{
+    use MakesGraphQLRequests;
+    use RefreshDatabase;
+    use RefreshesSchemaCache;
+    use UsesProtectedGraphqlEndpoint;
+
+    protected $platformAdmin;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+        Notify::spy(); // don't send any notifications
+        $this->bootRefreshesSchemaCache();
+        // Run necessary seeders
+        $this->seed(ClassificationSeeder::class);
+        $this->seed(RolePermissionSeeder::class);
+
+        $this->platformAdmin = User::factory()
+            ->asAdmin()
+            ->create([
+                'email' => 'platform-admin@test.com',
+                'sub' => 'platform-admin@test.com',
+
+                // The following properties make sure this user doesn't match certain test queries, skewing results.
+                'looking_for_english' => null,
+                'looking_for_french' => null,
+                'looking_for_bilingual' => null,
+                'accepted_operational_requirements' => null,
+                'location_preferences' => [],
+                'has_diploma' => false,
+                'position_duration' => [],
+                'is_gov_employee' => false,
+                'telephone' => null,
+                'first_name' => null,
+                'last_name' => null,
+            ]);
+    }
+
+    public function testCreateUserDefaultRoles()
+    {
+        $this->actingAs($this->platformAdmin, 'api')->graphQL(
+            /** @lang GraphQL */
+            '
+            mutation CreateUser($user: CreateUserInput!) {
+                createUser(user: $user) {
+                    firstName
+                    lastName
+                    email
+                    telephone
+                    preferredLang { value }
+                    preferredLanguageForInterview { value }
+                    preferredLanguageForExam { value }
+                }
+            }
+        ',
+            [
+                'user' => [
+                    'firstName' => 'Jane',
+                    'lastName' => 'Tester',
+                    'email' => 'jane@test.com',
+                    // If roles is not set, it should come back as empty array.
+                ],
+            ]
+        )->assertJson([
+            'data' => [
+                'createUser' => [
+                    'firstName' => 'Jane',
+                    'lastName' => 'Tester',
+                    'email' => 'jane@test.com',
+                    'telephone' => null,
+                    'preferredLang' => null,
+                    'preferredLanguageForInterview' => null,
+                    'preferredLanguageForExam' => null,
+                ],
+            ],
+        ]);
+        // Ensure user was saved
+        $this->assertDatabaseHas('users', ['email' => 'jane@test.com']);
+    }
+
+    public function testCreateUserAdminRole()
+    {
+        $this->actingAs($this->platformAdmin, 'api')->graphQL(
+            /** @lang GraphQL */
+            '
+            mutation CreateUser($user: CreateUserInput!) {
+                createUser(user: $user) {
+                    firstName
+                    lastName
+                    email
+                    telephone
+                    preferredLang { value }
+                    preferredLanguageForInterview { value }
+                    preferredLanguageForExam { value }
+                }
+            }
+        ',
+            [
+                'user' => [
+                    'firstName' => 'Jane',
+                    'lastName' => 'Tester',
+                    'email' => 'jane@test.com',
+                ],
+            ]
+        )->assertJson([
+            'data' => [
+                'createUser' => [
+                    'firstName' => 'Jane',
+                    'lastName' => 'Tester',
+                    'email' => 'jane@test.com',
+                    'telephone' => null,
+                    'preferredLang' => null,
+                    'preferredLanguageForInterview' => null,
+                    'preferredLanguageForExam' => null,
+                ],
+            ],
+        ]);
+        // Ensure user was saved
+        $this->assertDatabaseHas('users', ['email' => 'jane@test.com']);
+    }
+
+    public function testFilterByPoolCandidateStatuses(): void
+    {
+        // Get the ID of the base admin user
+        $user = User::All()->first();
+
+        // Create new pools and attach to new pool candidates.
+        $pool1 = Pool::factory()->create([
+            'user_id' => $user['id'],
+        ]);
+        $pool2 = Pool::factory()->create([
+            'user_id' => $user['id'],
+        ]);
+
+        PoolCandidate::factory()->count(5)->create([
+            'pool_id' => $pool1['id'],
+            'expiry_date' => config('constants.far_future_date'),
+            'pool_candidate_status' => PoolCandidateStatus::QUALIFIED_AVAILABLE->name,
+        ]);
+        PoolCandidate::factory()->count(4)->create([
+            'pool_id' => $pool1['id'],
+            'expiry_date' => config('constants.far_future_date'),
+            'pool_candidate_status' => PoolCandidateStatus::EXPIRED->name,
+        ]);
+        PoolCandidate::factory()->count(3)->create([
+            'pool_id' => $pool1['id'],
+            'expiry_date' => config('constants.far_future_date'),
+            'pool_candidate_status' => PoolCandidateStatus::PLACED_TERM->name,
+        ]);
+        PoolCandidate::factory()->count(2)->create([
+            'pool_id' => $pool2['id'],
+            'expiry_date' => config('constants.far_future_date'),
+            'pool_candidate_status' => null,
+        ]);
+        PoolCandidate::factory()->count(1)->create([
+            'pool_id' => $pool1['id'],
+            'expiry_date' => config('constants.far_future_date'),
+            'pool_candidate_status' => PoolCandidateStatus::REMOVED->name,
+        ]);
+
+        // Assert query with no pool filter will return all users, including unavailable
+        $this->actingAs($this->platformAdmin, 'api')->graphQL(
+            /** @lang GraphQL */
+            '
+            query getUsersPaginated($where: UserFilterInput) {
+                usersPaginated(where: $where) {
+                    paginatorInfo {
+                        total
+                    }
+                }
+            }
+        ',
+            [
+                'where' => [],
+            ]
+        )->assertJson([
+            'data' => [
+                'usersPaginated' => [
+                    'paginatorInfo' => [
+                        'total' => 16,
+                    ],
+                ],
+            ],
+        ]);
+
+        // Assert query with pool filter but no statuses filter
+        $this->actingAs($this->platformAdmin, 'api')->graphQL(
+            /** @lang GraphQL */
+            '
+            query getUsersPaginated($where: UserFilterInput) {
+                usersPaginated(where: $where) {
+                    paginatorInfo {
+                        total
+                    }
+                }
+            }
+        ',
+            [
+                'where' => [
+                    'poolFilters' => [
+                        [
+                            'poolId' => $pool1['id'],
+                        ],
+                    ],
+                ],
+            ]
+        )->assertJson([
+            'data' => [
+                'usersPaginated' => [
+                    'paginatorInfo' => [
+                        'total' => 13,
+                    ],
+                ],
+            ],
+        ]);
+
+        // Assert query with pool filter, only expired status
+        $this->actingAs($this->platformAdmin, 'api')->graphQL(
+            /** @lang GraphQL */
+            '
+            query getUsersPaginated($where: UserFilterInput) {
+                usersPaginated(where: $where) {
+                    paginatorInfo {
+                        total
+                    }
+                }
+            }
+        ',
+            [
+                'where' => [
+                    'poolFilters' => [
+                        [
+                            'poolId' => $pool1['id'],
+                            'statuses' => [PoolCandidateStatus::EXPIRED->name],
+                        ],
+                    ],
+                ],
+            ]
+        )->assertJson([
+            'data' => [
+                'usersPaginated' => [
+                    'paginatorInfo' => [
+                        'total' => 4,
+                    ],
+                ],
+            ],
+        ]);
+
+        // Assert query with pool filter, only removed status
+        $this->actingAs($this->platformAdmin, 'api')->graphQL(
+            /** @lang GraphQL */
+            '
+            query getUsersPaginated($where: UserFilterInput) {
+                usersPaginated(where: $where) {
+                    paginatorInfo {
+                        total
+                    }
+                }
+            }
+        ',
+            [
+                'where' => [
+                    'poolFilters' => [
+                        [
+                            'poolId' => $pool1['id'],
+                            'statuses' => [PoolCandidateStatus::REMOVED->name],
+                        ],
+                    ],
+                ],
+            ]
+        )->assertJson([
+            'data' => [
+                'usersPaginated' => [
+                    'paginatorInfo' => [
+                        'total' => 1,
+                    ],
+                ],
+            ],
+        ]);
+
+        // Assert query with pool filter, expired + removed + available statuses
+        $this->actingAs($this->platformAdmin, 'api')->graphQL(
+            /** @lang GraphQL */
+            '
+            query getUsersPaginated($where: UserFilterInput) {
+                usersPaginated(where: $where) {
+                    paginatorInfo {
+                        total
+                    }
+                }
+            }
+        ',
+            [
+                'where' => [
+                    'poolFilters' => [
+                        [
+                            'poolId' => $pool1['id'],
+                            'statuses' => [PoolCandidateStatus::QUALIFIED_AVAILABLE->name, PoolCandidateStatus::EXPIRED->name, PoolCandidateStatus::REMOVED->name],
+                        ],
+                    ],
+                ],
+            ]
+        )->assertJson([
+            'data' => [
+                'usersPaginated' => [
+                    'paginatorInfo' => [
+                        'total' => 10,
+                    ],
+                ],
+            ],
+        ]);
+
+        // Assert query with pool filter, empty array of statuses will return all candidates in pool.
+        $this->actingAs($this->platformAdmin, 'api')->graphQL(
+            /** @lang GraphQL */
+            '
+            query getUsersPaginated($where: UserFilterInput) {
+                usersPaginated(where: $where) {
+                    paginatorInfo {
+                        total
+                    }
+                }
+            }
+        ',
+            [
+                'where' => [
+                    'poolFilters' => [
+                        [
+                            'poolId' => $pool1['id'],
+                            'statuses' => [],
+                        ],
+                    ],
+                ],
+            ]
+        )->assertJson([
+            'data' => [
+                'usersPaginated' => [
+                    'paginatorInfo' => [
+                        'total' => 13,
+                    ],
+                ],
+            ],
+        ]);
+
+        // Assert query with unknown pool filter will return zero
+        $this->actingAs($this->platformAdmin, 'api')->graphQL(
+            /** @lang GraphQL */
+            '
+            query getUsersPaginated($where: UserFilterInput) {
+                usersPaginated(where: $where) {
+                    paginatorInfo {
+                        total
+                    }
+                }
+            }
+        ',
+            [
+                'where' => [
+                    'poolFilters' => [
+                        [
+                            'poolId' => '00000000-0000-0000-0000-000000000000',
+                            'statuses' => [PoolCandidateStatus::QUALIFIED_AVAILABLE->name],
+                        ],
+                    ],
+                ],
+            ]
+        )->assertJson([
+            'data' => [
+                'usersPaginated' => [
+                    'paginatorInfo' => [
+                        'total' => 0,
+                    ],
+                ],
+            ],
+        ]);
+    }
+}
